requirements.txt
# Core UI
streamlit>=1.33.0

# LangChain ecosystem
langchain>=0.2.0
langgraph>=0.0.45
langchain-groq>=0.1.3

# SerpAPI
google-search-results>=2.4.2

# Data validation
pydantic>=2.6.0

# Environment variable management
python-dotenv>=1.0.1

requests>=2.31.0